Add proof equivalence guest
