+++ blob-verification-methods/guest/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod blob_info;
+pub mod verify_blob;
