--- methods/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "blob-verification-methods"
-version = "0.1.0"
-edition = "2021"
-
-[build-dependencies]
-risc0-build = { workspace = true }
-
-[package.metadata.risc0]
-methods = ["guest"]
-
