--- methods/guest/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod blob_info;
-pub mod verify_blob;
