+++ proof-equivalence-methods/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "proof-equivalence-methods"
+version = "0.1.0"
+edition = "2021"
+
+[build-dependencies]
+risc0-build = { workspace = true }
+
+[package.metadata.risc0]
+methods = ["guest"]
+
