+++ proof-equivalence-methods/guest/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "proof-equivalence-guest"
+version = "0.1.0"
+edition = "2021"
+
+[workspace]
+
+[dependencies]
+alloy-primitives = { version = "0.8" }
+alloy-sol-types = { version = "0.8" }
+risc0-zkvm = { version = "1.2.0", default-features = false, features = ["std"] }
+serde = "1.0"
+rust-kzg-bn254-primitives = { git = "https://github.com/Layr-Labs/rust-kzg-bn254.git", branch = "master"}
+rust-kzg-bn254-prover = { git = "https://github.com/Layr-Labs/rust-kzg-bn254.git", branch = "master"}
+rust-kzg-bn254-verifier = { git = "https://github.com/Layr-Labs/rust-kzg-bn254.git", branch = "master"}
+ark-bn254 = "0.5"
+ark-ff = "0.5"
+num-bigint = "0.4.6"
+
+[patch.crates-io]
+# use optimized risc0 circuit
+crypto-bigint = { git = "https://github.com/risc0/RustCrypto-crypto-bigint", tag = "v0.5.5-risczero.0" }
+k256 = { git = "https://github.com/risc0/RustCrypto-elliptic-curves", tag = "k256/v0.13.3-risczero.0" }
+sha2 = { git = "https://github.com/risc0/RustCrypto-hashes", tag = "sha2-v0.10.8-risczero.0" }
